# [RFC] srepr displays assumptions used to create a Symbol