--- a/sympy/core/basic.py
+++ b/sympy/core/basic.py
@@ -45,7 +45,7 @@ class Basic(with_metaclass(ManagedProperties)):
     __slots__ = ['_mhash',              # hash value
                  '_args',               # arguments
                  '_assumptions',
-                 '_user_assumptions'
+                 '_user_assumptions',   # FIXME: _raw_assumptions or better?
                 ]
 
     # To be overridden with True in the appropriate subclasses