--- a/sympy/core/symbol.py
+++ b/sympy/core/symbol.py
@@ -64,7 +64,7 @@ class Symbol(AtomicExpr, Boolean):
             whose = '%s ' % obj.__name__ if obj else ''
             raise ValueError(
                 '%scommutativity must be True or False.' % whose)
-        # FIXME: what goes wrong if I drop this?
+        # NOTE: I want do sanitize without adding this, do it later elsewhere
         #assumptions['commutative'] = is_commutative
 
         # sanitize other assumptions so 1 -> True and 0 -> False
@@ -99,13 +99,18 @@ class Symbol(AtomicExpr, Boolean):
         False
 
         """
-        # FIXME: could make copy before sanitizing, avoid commutative=T/F.
-        # Easier to drop that from _sanitize as I did above...
+        # FIXME: could make copy before sanitizing: this avoids changes
+        # to the code related to strict commutative.  But easier to move
+        # commutative=T/F from _sanitize, otherwise need to do this in
+        # Dummy and Wild too.
         #acpy = assumptions.copy()
-        cls._sanitize(assumptions, cls)
-        obj = Symbol.__xnew_cached_(cls, name, **assumptions)
+        #cls._sanitize(assumptions, cls)
+        #obj = Symbol.__xnew_cached_(cls, name, **assumptions)
         #obj._user_assumptions = acpy
-        return obj
+        #return obj
+        cls._sanitize(assumptions, cls)
+        return Symbol.__xnew_cached_(cls, name, **assumptions)
+
 
     def __new_stage2__(cls, name, **assumptions):
         if not isinstance(name, string_types):
@@ -113,8 +118,10 @@ class Symbol(AtomicExpr, Boolean):
 
         obj = Expr.__new__(cls)
         obj.name = name
-        # FIXME: _raw_assumptions better name?
         obj._user_assumptions = assumptions.copy()
+        # be strict about commutativity
+        is_commutative = fuzzy_bool(assumptions.get('commutative', True))
+        assumptions['commutative'] = is_commutative
         obj._assumptions = StdFactKB(assumptions)
         return obj