Here's an attempt at #8042, to have those assumptions (and only those assumptions) with which a Symbol was created appear in its srepr.

E.g.,
````
x = Symbol('x', positive=1)
n = Symbol('n', even=True)
srepr(x)   % "Symbol('x', positive=True)"
srepr(n)   % "Symbol('n', even=True)"
````

@smichr: would you like to help?  This is inspired by #8025 and I'd appreciate your broader view!

I keep two dicts in private variables: a lightly sanitized version of the user input, and the `StdFactKB` (latter as it currently done).  This seems dangerous as I'm not convinced they could not diverge.  But its a starting point.  They can't change within a single immutable Symbol but as my `.as_dummy()` changes indicate, one has to be careful elsewhere.