@asmeurer 2015-01-12
Since this adds to `__slots__`, it's probably a good idea to do some performance and memory benchmarking (you can use memory_profiler to do memory benchmarking).