@smichr 2015-01-12
> @smichr: would you like to help?

Sorry, I'll have to pass on much help. But I would suggest that we aim for a minimal, canonical representation for the srepr form. Users aren't always aware of redundant assumptions. My suggestion is to *generate once (each time any change happens to assumptions occurs to SymPy master)* all possible combinations of assumptions and map to the outputs the minimal, canonical set of assumptions that gave that output. Some assumptions will be incompatible -- just ignore those since they can't occur in practice. If assumptions are ordered alphabetically and assumptions 1&3 or 2, 4&5 or 8&9 generate the same assumptions A, then 1&3 will be the minimal, canonical representation to use for assumptions A. Possible problems: 1) watch out for None values -- ignore them? 2) a set of assumptions may not fill the assumptions0 with all possible deduced values -- I guess that's ok if we generate all possible outcomes.