@asmeurer 2015-01-12
Using the assumptions given by the user seems like a good idea to me.