@cbm755 2015-01-13
@asmeurer:

**Timing**:  I did "./test assumptions" three times:
master: [29.73s, 29.61s, 29.21s]
this PR: [29.95s, 30.35s, 29.44s]

Doesn't look significantly slower.

**Memory**: `mprof run -T 0.01 cbm_memtest.py` where the script just does `from sympy import *`.

No significant difference.

master:
![mprof_master](https://cloud.githubusercontent.com/assets/818622/5720690/6c8f5360-9b1f-11e4-806c-a8313b5cde7e.png)
This PR:
![mprof_pr](https://cloud.githubusercontent.com/assets/818622/5720689/6c8f0fc2-9b1f-11e4-98c3-32160a325b05.png)

I've never benchmarked SymPy (or anything in Python) so if there is something more illuminating, just let me know.