@cbm755 2015-01-13
@smichr: I am also not convinced this PR is the right approach,  But what you describe is quite different.  I'll continue "as is" for now.