@cbm755 2015-01-13
An alternative approach might be to store the user-specified assumptions in a private variable in the  `StdFactKB`.  This sounds a bit nicer as we don't have the new slot added to Basic.  Sort of natural because the `StdFactKB` already knows about derived facts and it could also know from what it was initialized.  Would this be more palatable?

I will require some care regarding the "be strict about commutativty bit", unless I move that into `StdFactKB` as well...