@cbm755 2015-01-13
please also see #8819 for the "more palatable" version :)