@asmeurer 2015-01-13
I'm mostly concerned about the memory for something that creates a lot of Basic objects. I'm not sure if `assumptions` is the best test for that. Probably adding a single item to `__slots__` isn't that bad, but it's good to know.