@cbm755 2015-01-14
@asmeurer: ok.  Currently, I think I prefer #8812.  If you/others decide you prefer this approach (new `__slots__` in Basic), then we can benchmark further.  Or I can do it know if someone has a specific benchmark suggestion.