@asmeurer 2015-01-14
I prefer #8819 from a design perspective.