@asmeurer 2015-01-15
Let's try to get that one. Adding things to `__slots__` makes me uneasy.